services/embedding-service/tests/functional/pyproject.toml
[project]
name = "embedding-service-tests"
version = "0.1.0"
description = "Functional tests from embedding-service"
readme = "README.md"
requires-python = ">=3.12,<4.0"
dependencies = [
    "pytest>=8.3.4,<9.0.0",
    "aiohttp>=3.11.16,<4.0.0",
    "pytest-asyncio>=0.26.0,<1.0.0",
    "python-dotenv>=1.1.0,<2.0.0",
    "pydantic>=2.11.3,<3.0.0",
    "pydantic-settings>=2.9.1,<3.0.0",
    "numpy (>=2.3.1,<3.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "--asyncio-mode=auto -W ignore::DeprecationWarning"
